fix(topology): do not update the topology if the operation was cancelled while applying
The change must be cancelled only the operation cannot make progress. But
it might still happen that after it was cancelled, operation was applied
successfully and incorrectly apply the changes on the cancelled topology.
To prevent this, we now compare the expected topology with the current
topology before applying the change.