test: await between each setup command
The test update_job_timeout was flaky. Sometimes it would run the test
command and respond that it was rejected because the job could not be
found.

This is a race condition where zeebe has not processed the created
process instance far enough to have the job available by the time the
test command is run.

There is a pause between the setup commands and the test command, but
not between the individual setup commands.

It might be helpful to await in between each setup command, to give them
some time to be processed before the next is executed. However, it may
be wishful thinking that this resolves the flaky test fully.