feat(gateway): Zeebe Gateway fails to stream out activated jobs by not respecting the maxMessageSize
There were three options to solve original problem:
1) When mapping the broker response to the client response, do bookkeeping
on the response size to ensure the maxMessageSize.
2) Double the max inbound message size in the Zeebe Java Client by default.
3) Bring the WARNING: Stream Error exception into the control flow of the Gateway
so that the Gateway can react to it and potentially make the jobs re-activatable again.

I used mixed of option 1 and 3 for better UX. With option 2, it is still possible for
client users to override default maxInboundMessage which might result in the same issue.

In the ResponseMapper, we check if the gRPC response size of the activated job will
exceed the allowed maxMessageSize in the gateway. Afterward, we re-activate exceeding
jobs in RoundRobinActivateJobsHandler in a similar way that we do for failed jobs.

Using `getSerializedSize()` method shouldn't have any performance impact here since
the result of it will be cached to be used while actually returning the response.
See: https://protobuf.dev/reference/java/api-docs/com/google/protobuf/AbstractMessage.html#getSerializedSize--

(cherry picked from commit 58c0e20183ebc2faf03f38dd1109b21118a1f586)