fix(restore): validate partition count when restoring
Adds a new flag `validateConfig` that is enabled by default. When
enabled, we validate that the locally configured partition count
matches the number of partitions in the backup.

We allow setting `validateConfig` to false as an escape hatch for manual
recovery actions, for example to add or remove partitions.

(cherry picked from commit 4611c534af88ef36a409920ced73c41ac84b69c4)