refactor(engine): externalize checks for user tasks
The user task command precondition checker performs the common checks like tenant authorization.
Additional checks can be provided as a function upon initialization of the checker.
The additional checks are considered after the base checks have been performed.
If there are no additional checks or all of them succeed, the persisted record is returned.
Otherwise, the violation is returned.