fix(logstream): handle log reset while iterating though log entries
Iterating through the log is done by calling `hasNext` to check if there
is another entry to read, and `next` to actually retrieve it. Because
we don't hold a lock between `hasNext` and `next`, it can happen that
the log is reset between both calls. The `NoSuchElementException` thrown
by `next` would previously kill the partition which is unnecessary.

Instead, we now catch that exception and treat it as if `hasNext` had
returned false.

Closes #15721

(cherry picked from commit bbf6e45c5509ab87ad0e466d0eb4455028f3c8fb)