test(qa-integration-tests): put await to the correct place
refactor(gateway): use long for maxMessageSize

We would like to depend on spring framework only for
running application containers (e.g. gateway and broker).
Therefore, Spring's DataSize object is replaced with `long`.

refactor(gateway): rename maxMessageSize and add JavaDoc to toActivateJobsResponse

Since this maxResponseSize is specific to toActivateJobsResponse
method in the mapper, it is better if we add reasons for it to JavaDoc.

refactor(gateway): correct the actual size of returned response

For every ActivatedJob added to ActivateJobsResponse, 2 bytes are
added to the serialized size. While calculating the total response size,
we need to consider adding this to the total as well.

feat(gateway): handle if max message size exceeded after building the response

The statically added metadata size is removed because I discovered that it can
change based on the response data. See:
https://github.com/protocolbuffers/protobuf/blob/main/java/core/src/main/java/com/
google/protobuf/CodedOutputStream.java#L607
https://github.com/protocolbuffers/protobuf/blob/main/java/core/src/main/java/com/
google/protobuf/CodedOutputStream.java#L700

Instead, a final check is added to validate if the actual response size is exceeding
the max message size. I added the re-building to the end because building a gRPC response
is costly. Additionally, the case that built response size will exceed the max message size
is rare since metadata is expected to be <100 bytes long.

test(gateway): handle if max message size exceeded after building the response

A test is added to test the case where gRPC response size is bigger than
max message size when it is built. Numbers 5100 and 144 is used to actually trigger
the case. When we set jobVariableSize to 144, it produces ActivatedJob of size
1020 bytes. 5 jobs of 1020 bytes equals to 5100 bytes (the configured maxMessageSize).
But when we build the actual response it exceeds 5100 bytes and fall into our case.

(cherry picked from commit b2ad953cee58b9675b5f64a901f17a9ff412d937)