feat: allow to configure ES index suffix
Previously we always wrote the yyyy-MM-dd as suffix to every index.
This was hard-coded and not possible to configure/overwrite, which
can be seen as a limitation.

This commits makes it possible to configure the default suffix format
'yyyy-MM-dd' and overwrite it with something like: 'yyyy-MM-dd_HH'
which allows to create indices based on hours.

But it is also possible to create indices per day, if necessary and
not much data is expected.

This configures makes it possible to better leverage the ILM settings
provided by ES.