test(engine): verify user task action attribute
Ensures that the action attribute of user tasks is handled as expected in commands.
For all user task commands, the action has a default value relating to the original
name of the command. Custom values can be provided when creating the commands
and will be used in the events emitted as a result of the commands.