feat(engine): mark state with broker version that ran migrations
After running migration tasks (but still in the same transaction), we
now use the MIGRATION_STATE column family and write the current broker
version.
We don't make use of that information yet, but it will allow us to
detect incompatible states in the future.

The column family was (briefly) in use before to track which migration
tasks were executed. Since we only write a single key, there are no
conflicts with previously existing data.