test: limit recording exporter
This test case was taking 5+ seconds because one of the
RecordingExporter usages wasn't limited. I encountered this while
running all test cases in this package.

One way we can easily limit any record stream is if the process instance
is completed or terminated. In this case, the process instance is
canceled and thus terminated.

To use this with user task records, I needed to add an additional helper
method to filter the stream for only user task records.