fix(gateway): review fixes for gateway max message size assertion
* Removed unnecessary if check
* Updated tests to make sure all jobs are created. Otherwise, they might end up with
being flaky tests. Because in the old implementation, when we limit the stream to 5
records, it only throws an exception when the stream has no entries after 5 seconds.
It would wait 5 seconds and then continue if there is only a single job record.
Which can cause flakiness.

fix(gateway): review fixes for gateway max message size assertion

* Removed blocking join() calls
* Removed unused variables
* Added limit to job created assertion to speed up the test execution

test: assert jobs created before worker opened

This helps avoid multiple job activation roundtrips, because the worker
is only opened once all jobs are created.

Otherwise, there's a chance that the worker only receives some of the
jobs and has has to poll for more jobs until the final assertion is met.