fix(engine): limit iterator memory usage in Process and Decision migration state classes
In the previous implementation, one instance of RocksIterator was iterating through
the whole ColumnFamily. Since each instance of RocksIterator opens its transaction,
that led the previous implementation to occupy the memory equal to size of key-value
pairs.

RocksDB manages memory using malloc/free algorithm. In that implementation if there is
enough memory left in the total allocated memory, it keeps freed memory unused, because
it expects it to be used again by the same object. For that reason, when RocksIterator
occupies too much memory, there is a possibility that after it is done, the free memory
might not be allocated for other threads.

In the updated implementation, the memory usage per iterator is limited to a fixed size.
This will provide OS to re-use (or at least keep less sized non-used memory) occupied memory
when it is no longer used by another iterator.

The fixed limit is chosen as default 50MB. It is because, 50MB appeared to be a sweet spot
while testing the fix. When I use 100MB, it was almost doubling the memory usage and when
I use 25MB, it only increased the migration time while not improving on the memory usage.

(cherry picked from commit 4a0de147ec58f27b733e44f1e01af7200b27c481)
(cherry picked from commit 323274715a1e3c623c1d95b97f10fe1d50ffbcc4)