fix(engine): remove transaction from running migrations
For the same reason as mentioned in the previous commit, we should keep the memory
usage of each transaction low. Otherwise, the data retrieved will be kept in memory
and after its usage, the freed memory might not be reused.

If we do not remove this transaction, we will not benefit from memory limiting each
`columnFamily.foreach` loop because it will keep changes uncommitted. Therefore, it
will cause occupying the whole migration data in memory until it is committed.

(cherry picked from commit d94a32b87da923db98be775c127f63277aa0303a)