fix(engine): update user task state correctly
The user task record handed into an applier needs to extract all required data
before fetching the user task stored in the state. Otherwise, the data of the
record handed into the applier is adjusted and the target values are lost.

For the ASSIGNED applier, the assignee is extracted. For the UPDATED applier,
the changed attributes and the attributes tracking changes are extracted.

The user task state is reduced to updating column families. Previously, it
took care of more complex operations like `update` that also fetches the state,
updates the stored record, and stores it back into the state. This code now
resides in the appropriate appliers.