refactor(protocol): column families impl enumvalue
This doesn't change any behavior. We introduce a new EnumValue interface
that can be implemented by Enums to avoid using ordinal as the value
associated with each enum value.

We also implement this in the ZbColumnFamilies enum as this suffers from
bugs caused by ordinal re-ordering. While this doesn't yet move
ZbColumnFamilies away from using an ordinal (note the implementation),
it already prepares the enum for moving to a hardcoded number system.

We'll add this number system in a later commit.