refactor(protocol): column families impl enumvalue
This doesn't change any behavior. We introduce a new EnumValue interface
that can be implemented by Enums to avoid using ordinal as the value
associated with each enum value.

We also implement this in the ZbColumnFamilies enum as this suffers from
bugs caused by ordinal re-ordering. While this doesn't yet move
ZbColumnFamilies away from using an ordinal (note the implementation),
it already prepares the enum for moving to a hardcoded number system.

We'll add this number system in a later commit.

(cherry picked from commit 281c3219153f52e1e66ac9191ba83cc8ee7655f3)

The following column families are not yet supported on this version:
  SIGNAL_SUBSCRIPTION_BY_NAME_AND_KEY,
  SIGNAL_SUBSCRIPTION_BY_KEY_AND_NAME,

  USER_TASKS,
  USER_TASK_STATES,
  COMPENSATION_SUBSCRIPTION

Conflicts:
  protocol/src/main/java/io/camunda/zeebe/protocol/ZbColumnFamilies.java