refactor(protocol): column families impl enumvalue
This doesn't change any behavior. We introduce a new EnumValue interface
that can be implemented by Enums to avoid using ordinal as the value
associated with each enum value.

We also implement this in the ZbColumnFamilies enum as this suffers from
bugs caused by ordinal re-ordering. While this doesn't yet move
ZbColumnFamilies away from using an ordinal (note the implementation),
it already prepares the enum for moving to a hardcoded number system.

We'll add this number system in a later commit.

(cherry picked from commit 281c3219153f52e1e66ac9191ba83cc8ee7655f3)

There are conflicts because ZbColumnFamilies was changed but that file
still exists in engine on 8.1. It was moved to protocol in 8.2.

To resolve the issue, I copied the relevant changes to the old
ZbColumnFamilies file.

Conflicts:
  protocol/src/main/java/io/camunda/zeebe/protocol/ZbColumnFamilies.java