refactor(zb-db): expect cf type with EnumValue
We're introducing EnumValue as a way to associate hardcoded values with
the ZbColumnFamilies rather than the currently used ordinals.

zb-db doesn't depend on ZbColumnFamilies directly, so it can function
regardless of the data that is stored in it. This makes sense, because
the data schema is decided by modules like engine (in the state
package).

To allow using EnumValue's .getValue() in subsequent commits inside the
zb-db module, we need to clarify some structure for the ColumnFamilyType
generics: it must be an enum that implements EnumValue.

We can achieve that by combining some generics. First, we say that the
type extends an Enum that itself extends EnumValue.

However, then we don't have access to the .getValue() method yet. To add
that, we also have to say that the type extends EnumValue itself. This
combination can be achieved with Intersection Types (A & B).

Note that for the test column family enums, it doesn't matter how the
getValue method is implemented. We can use .ordinal() in the test
classes, because they are not used in production code. While,
DefaultColumnFamily is also not used in production code, it does exist
in `main` and could be used in production later. Therefore, I've
implemented that getValue with an actual hardcoded value.

(cherry picked from commit 7794eae6ce70c1e40745a3d12968f756c2ef1cec)

The following file does not exist on this version, but has changed in
the original commit.

Conflicts:
  zb-db/src/main/java/io/camunda/zeebe/db/impl/rocksdb/SnapshotOnlyDb.java