fix(protocol): provide default values in DMN deployment records
DMN resources that pass model validation but fail to deploy due to missing data,
e.g. an empty decision name, can lead to unwritable data on the stream.
A rejection for the deployment CREATE command can thus not be written due to MsgPack
exceptions. This can run the message writer into an infinie loop of trying to write.

Using proper defaults in the DMN deployment records lets the writer successfully write
the deployment CREATE command rejection.