test(protocol): avoid reuse of enum values in cf
This adds two tests.

The main test is shouldNotReuseEnumValues. It ensures that the
getValue() method produces a unique value for each value in the enum.
It does not test determinism, but that is implied for this method.

The current implementation uses the ordinal() method. To reimplement it,
I've added a temporary test that we can remove after re-implementing the
getValue() method. It should pass after re-implementation to ensure we
don't accidentally reorder the values now.

(cherry picked from commit e415790bf753be809e6ba510946cbf78ba0e6295)

There wasn't necessarily a conflict in this commit, but ZbColumnFamilies
doesn't exist in the protocol module on 8.1. Instead, it exists in
engine.

So, I've moved the test class ZbColumnFamiliesTest over to the engine
module. This will likely cause some conflicts later.