test(protocol): avoid reuse of enum values in cf
This adds two tests.

The main test is shouldNotReuseEnumValues. It ensures that the
getValue() method produces a unique value for each value in the enum.
It does not test determinism, but that is implied for this method.

The current implementation uses the ordinal() method. To reimplement it,
I've added a temporary test that we can remove after re-implementing the
getValue() method. It should pass after re-implementation to ensure we
don't accidentally reorder the values now.