refactor: use hardcoded column family enum values
While this is a refactor: it does not change behavior, it strengthens
our ability to add and backport new column families by hardcoding the
associated value.

A test exists that verifies that all hardcoded values are added using
the exact same value as the ordinal currently has.

Another test exists that verifies that no hardcoded value is reused.
This one is especially important, as it protects against accidental
problems while automatically backporting.

After this change, values don't necessarily need to be added to the
bottom anymore, as long as they receive a unique value.

(cherry picked from commit 91182697fa264022fed36b7c69ab61968e01dd1f)

Multiple conflicts existed. As the ZbColumnFamilies exists in protocol
on newer version (8.2+) but is still in engine module on 8.1, I had to
manually copy the changes.

I did this by copying the latest version of the file, and working back
per entry:
- reverting renames for deprecation
- removing column families that are not yet supported on 8.1
- adjusting the value of the three faulty column family ordinals:
  - MESSAGE_STATS 48 should be 54
  - PROCESS_INSTANCE_KEY_BY_DEFINITION_KEY 49 should be 55
  - MIGRATIONS_STATE 50 should be 56

We'll need to change these values in an upcoming pull request to fix the
bug 15804 on 8.1.

Conflicts:
  protocol/src/main/java/io/camunda/zeebe/protocol/ZbColumnFamilies.java