refactor: use hardcoded column family enum values
While this is a refactor: it does not change behavior, it strengthens
our ability to add and backport new column families by hardcoding the
associated value.

A test exists that verifies that all hardcoded values are added using
the exact same value as the ordinal currently has.

Another test exists that verifies that no hardcoded value is reused.
This one is especially important, as it protects against accidental
problems while automatically backporting.

After this change, values don't necessarily need to be added to the
bottom anymore, as long as they receive a unique value.