fix(zb-db): use static value in foreignkey lookup
As database entries are stored with the column family's value prefixed
to the key, so too do we look them up with that value when checking
foreign key constraints.

We assume that the foreignKey.columnFamily() returns an instance of
EnumValue so we can call the .getValue() on it. This is safe to do
because the column family must anyways be an instance extending Enum<?
extends EnumValue>.

Alternatively, we could've avoided this cast if we would make the
foreign key understand that the column family type also actually
implements EnumValue, using an intersection type. However, that would've
required adding this intersection type as an additional generic on the
DbForeignKey<DbKey, CFType>. As this class is used often in our state
classes, it would deteriorate the readability of those classes, as we
currently can refer to it as DbForeignKey<DbKey>. This is why I went for
casting rather than more generics.