feat(workflows): Increase maxInstanceDuration in Weekly Benchmarks
There was a test failure in one of our weekly e2e tests. From the logs and metrics
it looks like a starter applies a backoff after 10 minutes but since the instance
duration is set to 15 minutes by default, the test fails before the processing catches up.

For that reason, we would like to increase the instance duration to 40m to see if it is
actually catching up with the processing.