test: map to different element id
Previously, we've had an issue with migrating element instances to
elements with a different id which was not caught in tests. While it
shouldn't be a problem in this test case, it's useful to make it a
habbit in most test cases to migrate to element with different id.

The case where we migrate to element with the same id, is already
covered elsewhere.