test: ensure migrated user task event without vars
We want to avoid surpassing the max message size. As the task variables
will not change on migration (at least at this time), we can simply
write them as null.

Note that when deserializing the `null` value, we see an empty map on
the recorded stream. I believe that this is due to the DocumentProperty.