fix: correct skewew column families
Redefines the column families that were wrong compared to 8.0 and all
future versions.

This also recovers the data from these column families.

Note that the corrector for column family 50 was simplified, because
SignalSubscriptions don't exist on 8.1 (or before). So no entries in
this column family are correctly in there. All must be entries that
belong in the MIGRATIONS_STATE CF. So we can move all of them.

Without this simplification we'd have to introduce the
SignalSubscriptionRecord and related classes just to deal with these
entries that cannot exist here anyways.