feat: reject migration on inconsistent state
This adds a new exception SafetyCheckFailedException that can be thrown
in the ProcessInstanceMigrationMigrateProcessor whenever a safety check
fails. Typically, this means a bug is present.

When thrown, we log the error with stacktrace and reject the migration
command. As this is likely a bug, we instruct the user to report it as a
bug.

We also already use the exception by throwing it in case we're expecting
to find a job or a user task in the state, but don't find it.

Lastly, I've added a new logger to represent processing on the engine.
Processing is a package of the engine that contains everything related
to processing.