feat(raft): add sender id to RaftRequest
Most requests have info about the sender already. But each request has a different name
for it. For uniform access to this info, we add a get method to the RaftRequest interface.

ReConfigure request didn't have a sender. So we added it. Adding a new field is
backward compatible with the kryo serializer we use.