refactor(topology): differentiate between appliers that changes ClusterTopology or only local MemberState
Previously operations could only change the state of the local member. With the force
reconfigure, the member that applies this operation must also remove the partition from
the members that will be removed.

The interface for TopologyChangeApplier didn't allow to change other than the local member.
To allow force reconfigure to change other members, we refactor the interface. The new
interface differentiate between the appliers that can change only the local member
(MemberOperationApplier) and the appliers that can change other members
(ClusterOperationApplier). The previous appliers implements MemberOperationApplier.