refactor: remove gateway-protocol-rest module
Remove the gateway-protocol-rest module. The goal of the module was to
generate REST API model code that would have been reused in the REST API
controllers and the REST API client.

However, the code in the generated model files varies depending on the
OpenAPI generator that is used. For the Spring Boot-based REST API
server that will be implemented in Zeebe, the spring OpenAPI generator
is more appropriate since it is a server-type generator.

On the other hand, the OpenAPI java generator is a client-type generator
and generates significantly different code requiring the specification
of client-side libraries (like feign or okhttp) which are not relevant
for the server side and limit the usage of this module for any users
that would like to implement their own Zeebe REST API client with their
HTTP client selection.