refactor(gateway-rest): add openapi model code generation
Move the OpenAPI spring generator to the gateway-rest module. This is
done since the generated model code varies depending on the application
that will use it. Since in the Zeebe Gateway, a Spring Boot-based REST
API will be used, we need to generate appropriate models. As a result,
the code generation for the REST API models is bundled in the same
module as the REST API controllers.