fix: open db with cf handles
Somehow, the default column family could no longer be retrieved with
.getDefaultColumnFamilyHandle method. It's unclear why this method
doesn't work anymore, but it returns null now.

This applies a different way to retrieve the default column family
handle by using a different method to open the db. This method allows
passing the database options separate from the column family
descriptors (name + cf options), so you can open the db for multiple
column families. See RocksJava Basics.

The method adds the handles for the column families to the passed in
column family handles list parameter. As we're only interested in the
default column family, we can simply expect a single handle.

Note that this method requires DbOptions rather than Options (which
combines the DbOptions and ColumnFamilyOption). The order of opening and
closing is important. See RocksJava Basics.

RocksJava Basics:
https://github.com/facebook/rocksdb/wiki/RocksJava-Basics#opening-a-database-with-column-families