fix(raft): anoint has priority over proactive leadership transfer
We now have two kinds of leadership transfer.
1. When a follower is annointed, it sends a transfer request to the
leader
2. When the leader role closes, it sends a transfer request to a
follower

This fixes an issue with annointing followers. When the leader responds
to this, it transitions to follower before replying to the request.
The expectation was that the follower sending the request then becomes
candidate and leader. Without this fix however, the transition to
follower would also cause a proactive transfer that could make any other
follower a candidate and leader.