refactor(protocol): clean up breaking changes for revapi
Replacing the candidate properties in user tasks is a short-term change
that goes into the `ignored-changes.json` only relevant for updating from
8.4 to 8.5.

Remove the changes from the long-term changes in `revapi.json` that are always considered.
Add an `ignored-changes.json` for the next release and add the changes there.
Improve the justifications for replacing the candidate properties in user tasks.