test: don't assert commands
Most of our tests should focus on asserting events rather than commands.
Commands can be rejected and only represent what we want to happen, but
it doesn't mean that those things actually happen.

In this case, the Timer TRIGGER command could be appended multiple times
by the checker if the checker is faster than the stream processor.

If that happens, this test failed, because there would be two TRIGGER
commands encountered. The TRIGGERED event would not be encountered
because the limit is set to 2. Note that the second TRIGGER command
would anyways be rejected, so it doesn't have any effect.

(cherry picked from commit e332eb7c3d30c8c4cfdffc2cd3dac6051f460a47)