test: only assert events
Like the test changed in the previous commit, this test case was also
asserting events and commands in a certain order unnecessarily. This
test case doesn't have to assert commands, because it simply wants to
verify that the process instance is created as expected. That we use a
command in between the activation and completion of the start event is
irrelevant as its an implementation detail.