test(qa): add tests to verify DbFormState cache behaviour
While putting the form to the cache during the retrieval for linking, we skip copying
a re-used form record object. That causes an issue that the cache contains a reference
to a singular dbPersistentForm object and that is modified with a new form deployment.
Therefore, a test is added to verify the wrong behaviour.

**Steps of the test case**
- Deploy a form with id formId1
- Deploy a process (id=process1) that contains a user task which has a link to form formId1
- Restart the broker
    - Take a snapshot
    - Stop the broker
    - Start broker from the last snapshot
- Create a process instance of process1 (the form linking is correct)
- Deploy another form with id formId2
- Create a process instance of process1 (the form linking is wrong) => it will retrieve the key for formId2