fix(engine): fix DbFormState cache behaviour
While putting the form to the cache during the retrieval for linking, we skip copying a re-used form record object.
Then, when we insert another form, we modify that object. That means, we modify the object that we put to the cache.
Afterward, when we query formId1 from the cache, it returns the wrong form object (formId2's object in our case).
It happens only after broker restart because the cache is cleaned up during restart. Therefore, the correct form we
put to the cache during the creation of form is gone.

During the form linking, the retrieved form is first copied to fix the incorrect behaviour.