feat(gateway-rest): add a rest controller for the user task api
Define a REST API Controller class that will encapsulate the
REST API endpoints of the user task API.

Initially, only the POST /user-tasks/userTaskKey/completion endpoint is added.

Instead of synchronously waiting for the broker response, we return futures in the
controller to let the REST infrastructure take over handling waiting for future completions.

The controller handles the error use cases consistently for broker rejections and errors.
Details of the rejections or errors are returned in the PrblemDetail payload.

A generalized error mapper for broker invocations in the REST API yields ProblemDetail objects.
It can handle broker errors, rejections, and exceptions, depending on what the broker returned.
The user task controller uses the mapper to build responses from the problem details.

Rejection handling is defined in the controller since rejections can have different meanings for
each controller. The controller hands over the rejection handling routine to the error mapper.

Request mapping is centralized in a utility class. Controllers map REST request to broker requests
using this utility class.