test: surface bug by changing id
This test case was not surfacing a bug due to usage of a specific
user-defined string as the decisionRequirementsId. This meant that the
correct data did not trigger the false positive where it would be
attempted to be moved.

(cherry picked from commit 7bdbf0f648ce6a7bfc5bb42dc378cadd56a9a7b3)