fix: move data that is not a drg
Previously, this code tried to move the data if it did not match the
processInstanceKeyByProcessDefinitionKey but actually it should move it
when it does match that.

another way to look at it, is that it should leave entries that match
the decisionRequirementsIdAndVersion key should be left alone because
they belong in this column family. Only entries that cannot be matched
to the decisionRequirementsIdAndVersion should be moved.