fix(engine): check state version for compatibility
If the current state is marked as migrated by some semantic version that
is higher than the current version, we know that this comes from a
snapshot taken by a newer broker. In that case we can't just use the
state because it may contain data that we can't work with.
Here we just throw an exception which results in the migration step
failing. As a result, the `ZeebePartition` does not start up correctly,
preventing incorrect migrations, processing or replay.