feat(topology): start coordinator services in all brokers
Previously it was only started in broker 0. But now we want
to allow all brokers to be the coordinator. Note that at a time
only one broker can be the coordinator which is selected based
on the lowest memberId. But when the broker with the lowest memberId is removed, the next broker should be able to handle all topology requests.