feat(engine): remove userTaskKey from instance on complete
When a user task is completed, its reference from the element instance needs to be removed.
This aligns the user task completiong with the job completion and consolidates how lifecycles
of sub-elements of the element instance work.
This makes building features on top like migration a lot easier since no extra checks have
to be considered for user tasks specifically.