test: should close msg sub when unmapped
When migrating a process instance that is currently subscribed to a
message event we must unsubscribe from that message event if the
migration does not require subscribing to that same message again in the
target process. I.e. when the message subscription is not mapped.

This can happen for two cases currently:
- message boundary event
- message event sub process

The other cases are not yet supported:
- receive task
- message catch event
- message catch event attached to event based gateway