refactor: allow distribution without typed record
The command distribution only needs a value type, intent, and value.
Previously, we could simply pass in the typed record when using the
command distribution behavior. For example, for distributing a broadcast
signal command, we'd be distributing the typed record (command) that was
currently being processed.

However, for migration of message subscriptions, we're not currently
processing a message subscription command. And since the TypedRecord
implementations mainly exist in the stream processor, it's good to avoid
instantiating them in the engine.

This commit simply adds an overloaded method for distributeCommand that
doesn't take the TypedRecord as a command, but breaks it up into the
smaller parts: value type, intent, and value.