refactor(gateway-rest): validate user task update request
Reject update requests for user tasks that do neither contain an action nor changes.
Such requests will be answered with a 400, stating the missing update data.

The controller and RequestMapper are refactored to use Either types to account for
validation errors and proper broker requests respectively. This allows for code reduction
and more concise controller code. Plus, request validation becomes a part of request mapping
itself, which is more consistent and easier to track in a single place.