feat(client/java): add maven dependecies for rest api support
Add Apache HttpClient 5 as a dependency to the Zeebe Java client for
the purposes of implementing support for the Zeebe REST API.

Add the openapi-generator-maven-plugin to the Zeebe Java client to
generate REST API model classes from the Zeebe OpenAPI
specification. The models will be used by the http
client to map request and response objects.

The JSR305 dependency is required by the OpenAPI spec generator
in order to support nullable annotations to the generated REST
API models (for more details see
https://github.com/OpenAPITools/openapi-generator/issues/7177).