test: unsub msg boundary on migration
Verifies a very basic case: we unsubscribe from a message boundary event
when we migrate an active element that is subscribed to it.

In the example case, we have a user task A that is migrated to a user
task B. User task A is subscribed to a message because of its boundary
message event. On migration, there is no mapping instruction provided
for the boundary event, so the message subscription is deleted.