test: sub msg boundary on migration
Verifies a very basic case: we subscribe to a message boundary event
when we migrate an active element to an element with that message
boundary event.

In the example case, we have a user task A that is migrated to a user
task B. User task B becomes subscribed to a message because of its
boundar message event. On migration, there is no mapping instruction
provided for the boundary event, so the message subscription is created.