feat(client/java): add communication api config property
Add a configuration property in the Java client to set the
default communication API for the client commants. Users
can pick between REST and GRPC, with REST being the default
if the property is not explicitly set.

Furthermore, a new CommandWithCommunicationApi interface is
added to support setting the communication API explicitly
implementation is stable and implemented across all of the
client commands that would support multiple communication
APIs.