feat(client/java): support communication api setting in topology request
Implement the CommandWithCommunicationApiStep in the TopologyRequest
command. The implementation of this interface will allow us to
explicitly pick between REST and GRPC when a TopologyRequest is made
through the Java client.

With this commit, the command will only use GRPC as an http client
hasn't been implemented yet. This will change with the upcoming
commits.