feat: only subscribe to new messages
This changes the filter for subscribeToMessageEvents to allow checking
that we're only subscribing to new messages. We can skip subscribing to
messages that we just subscribed to as part of this migration.

This requires subscribing to events before unsubscribing from events,
because otherwise we don't know which messages we were already
subscribed to before the migration.

When subscribing to new message events, we can collect them in a list.
This list we can then use as part of the filter for unsubscribing.