refactor: focus on boundary events not event types
This clarifies better what the code is doing: It's looking for boundary
events of certain types, as only specific types are allowed. Those that
are not allowed result in a rejection of the migration, i.e. they are
rejected boundary events. The type can be inferred by the reader,
because the stream's type is left unchanged, it's just filtered and
collected.

This introduces a tiny duplication (getEventType is called twice), at
the benefit of much simpler to follow logic. We look at all boundary
events, and those where the type is not allowed are rejected. At the
end, those types are collected in a way that fits the error reason.