test(zeebe): Verify variable propagation for compensation handler
Add a test case to verify that the variables of a compensation handler are propagated to the expected scope.

If the compensation throw event is inside a subprocess then the variables of the compensation handler should override
local variables of the subprocess, or create new variables in the root process instance scope.