refactor(engine): apply review suggestions
The assertions for the incident record with `RESOLVED` intent is not added
in this commit. I will add them in the next commit with the `IncidentIntent.
MIGRATED` event is applied in the state. When the applier is added, the
incident record with `RESOLVED` intent is expected to contain updated
`processDefinitionKey`, `bpmnProcessId` and `elementId`.

refactor(engine): remove unnecessary assertion

The expectation is already verified by the `resolve()` method itself.
No need to add an extra assertion.

feat(engine): update incident migrated event in the state

I added new assertions to the test cases that we make sure the resolved
incident contains the updated values after the migration. It is important
because exported record consumers should see an updated incident record
after the migration.

In the incident state, I introduced a new `updateIncidentRecord` method to
better represent its implementation and call it in `migrateIncident` method.

I also added "updated" metrics for later use in the metrics if needed.

refactor(engine): remove metric for update incident migrated event

The incident updated metric is removed because we would like to add migration metrics
in a separate issue as a whole.

refactor(engine): remove unnecessary column family update for incident migrated event

Since we do not store the record value in `jobIncidentColumnFamily` and
`processInstanceIncidentColumnFamily` we do not need to update these column families.